print size_t by casting to unsigned long
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@14719 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ appl/test/uu_client.c
@@ -143,9 +143,9 @@ proto (int sock, const char *hostname, const char *service)
 
 	authz = ticket->ticket.authorization_data;
 	for (i = 0; i < authz->len; i++) {
-	    printf("\ttype %d, length %d\n",
+	    printf("\ttype %d, length %lu\n",
 		   authz->val[i].ad_type,
-		   authz->val[i].ad_data.length);
+		   (unsigned long)authz->val[i].ad_data.length);
 	}
     }
